+++ src/main/java/de/alpharogroup/checksum/ByteArrayChecksumExtensions.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
  * including without limitation the rights to use, copy, modify, merge, publish, distribute,
  * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all copies or
  * substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
  * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
@@ -20,34 +20,27 @@
  */
 package de.alpharogroup.checksum;
 
+import de.alpharogroup.crypto.algorithm.Algorithm;
+
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.zip.Adler32;
 import java.util.zip.CRC32;
 import java.util.zip.Checksum;
 
-import de.alpharogroup.crypto.algorithm.Algorithm;
-
 /**
  * The class {@link ByteArrayChecksumExtensions} is a utility class for computing checksum from byte
  * arrays
  *
- * @version 1.0
  * @author Asterios Raptis
+ * @version 1.0
  */
 public final class ByteArrayChecksumExtensions
 {
-	/** The constant HEXADECIMAL_CHARACTER_CLASS. */
-	private static final String HEXADECIMAL_CHARACTER_CLASS = "[a-fA-F0-9]";
 
-	/** The constant REGEX_VALIDATION_MD5. */
-	private static final String REGEX_VALIDATION_MD5 = HEXADECIMAL_CHARACTER_CLASS + "{32}";
-
-	/** The constant REGEX_VALIDATION_SHA1. */
-	private static final String REGEX_VALIDATION_SHA1 = HEXADECIMAL_CHARACTER_CLASS + "{40}";
-
-	/** The constant REGEX_VALIDATION_SHA512. */
-	private static final String REGEX_VALIDATION_SHA512 = HEXADECIMAL_CHARACTER_CLASS + "{128}";
+	private ByteArrayChecksumExtensions()
+	{
+	}
 
 	/**
 	 * Gets the checksum from the given byte arrays with the given algorithm
@@ -191,42 +184,6 @@ public final class ByteArrayChecksumExtensions
 		return checksum.getValue();
 	}
 
-	/**
-	 * Checks if the given value matches a MD5 value pattern.
-	 *
-	 * @param value
-	 *            the value
-	 * @return true, the given value matches a MD5 value otherwise false
-	 */
-	public static boolean matchesMD5(final String value)
-	{
-		return value.matches(REGEX_VALIDATION_MD5);
-	}
-
-	/**
-	 * Checks if the given value matches a SHA1 value pattern.
-	 *
-	 * @param value
-	 *            the value
-	 * @return true, the given value matches a SHA1 value otherwise false
-	 */
-	public static boolean matchesSHA1(final String value)
-	{
-		return value.matches(REGEX_VALIDATION_SHA1);
-	}
-
-	/**
-	 * Checks if the given value matches a SHA512 value pattern.
-	 *
-	 * @param value
-	 *            the value
-	 * @return true, the given value matches a SHA1 value otherwise false
-	 */
-	public static boolean matchesSHA512(final String value)
-	{
-		return value.matches(REGEX_VALIDATION_SHA512);
-	}
-
 	private static byte[] toByteArray(Byte[] bytes)
 	{
 		byte[] byteArray = new byte[bytes.length];
@@ -235,10 +192,6 @@ public final class ByteArrayChecksumExtensions
 			byteArray[i] = bytes[i];
 		}
 		return byteArray;
-	}
-
-	private ByteArrayChecksumExtensions()
-	{
 	}
 
 }