+++ src/main/java/de/alpharogroup/checksum/FileChecksumExtensions.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
  * including without limitation the rights to use, copy, modify, merge, publish, distribute,
  * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all copies or
  * substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
  * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
@@ -19,6 +19,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package de.alpharogroup.checksum;
+
+import de.alpharogroup.crypto.algorithm.Algorithm;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -30,17 +32,19 @@ import java.util.zip.Adler32;
 import java.util.zip.CRC32;
 import java.util.zip.CheckedInputStream;
 
-import de.alpharogroup.crypto.algorithm.Algorithm;
-
 /**
  * The class {@link FileChecksumExtensions} is a utility class for computing checksum from files and
  * byte arrays.
  *
- * @version 1.0
  * @author Asterios Raptis
+ * @version 1.0
  */
 public final class FileChecksumExtensions
 {
+	private FileChecksumExtensions()
+	{
+	}
+
 	/**
 	 * Gets the checksum from the given file with an instance of the given algorithm.
 	 *
@@ -84,9 +88,9 @@ public final class FileChecksumExtensions
 	public static long getChecksum(final File file, final boolean crc)
 		throws FileNotFoundException, IOException
 	{
-		try (CheckedInputStream cis = crc
-			? new CheckedInputStream(new FileInputStream(file), new CRC32())
-			: new CheckedInputStream(new FileInputStream(file), new Adler32()))
+		try (CheckedInputStream cis = crc ?
+			new CheckedInputStream(new FileInputStream(file), new CRC32()) :
+			new CheckedInputStream(new FileInputStream(file), new Adler32()))
 		{
 			final int length = (int)file.length();
 			final byte[] buffer = new byte[length];
@@ -118,8 +122,8 @@ public final class FileChecksumExtensions
 	public static String getChecksum(final File file, final String algorithm)
 		throws NoSuchAlgorithmException, IOException
 	{
-		return ByteArrayChecksumExtensions.getChecksum(Files.readAllBytes(file.toPath()),
-			algorithm);
+		return ByteArrayChecksumExtensions
+			.getChecksum(Files.readAllBytes(file.toPath()), algorithm);
 	}
 
 	/**
@@ -148,10 +152,6 @@ public final class FileChecksumExtensions
 	public static long getCheckSumCRC32(final File file) throws IOException
 	{
 		return ByteArrayChecksumExtensions.getCheckSumCRC32(Files.readAllBytes(file.toPath()));
-	}
-
-	private FileChecksumExtensions()
-	{
 	}
 
 }